refact: `BaseMatrix`
