+++ src/poly/dcrt/matrix/mod.rs
@@ -1,5 +1,5 @@
+pub mod base;
 pub mod dcrt_poly;
-pub mod generic_matrix;
 pub mod i64;
 
 pub use dcrt_poly::DCRTPolyMatrix;
